# Project Management/src/projectManagement.API/wwwroot/openapi-original.json
{
  "openapi" : "3.0.3",
  "info" : {
    "description" : "Project Management API",
    "title" : "Project Management API",
    "version" : "1.0"
  },
  "servers" : [ {
    "url" : "https://user.server.test/v1"
  } ],
...
  "paths" : {
    "/Users" : {
      "get" : {
        "description" : "Retrieve a list of users",
        "operationId" : "getUsers",
        "parameters" : [ {
          "explode" : true,
          "in" : "query",
          "name" : "pageIndex",
          "required" : false,
          "schema" : {
            "default" : 1,
            "format" : "int64",
            "type" : "integer"
          },
          "style" : "form"
        }, {
          "explode" : true,
          "in" : "query",
          "name" : "pageSize",
          "required" : false,
          "schema" : {
            "default" : 10,
            "format" : "int64",
            "type" : "integer"
          },
          "style" : "form"
        }, {
          "explode" : true,
          "in" : "query",
          "name" : "totalCount",
          "required" : false,
          "schema" : {
            "default" : 1,
            "format" : "int64",
            "type" : "integer"
          },
          "style" : "form"
        }, {
          "explode" : true,
          "in" : "query",
          "name" : "pageNumber",
          "required" : false,
          "schema" : {
            "default" : 1,
            "format" : "int64",
            "type" : "integer"
          },
          "style" : "form"
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/getUsers_200_response"
                }
              }
            },
            "description" : "OK"
          },
          "default" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/error"
                }
              }
            },
            "description" : "Unexpected error"
          }
        },
        "summary" : "Get Users",
        "tags" : [ "User" ]
      },
      "post" : {
        "operationId" : "createUser",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/User"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "201" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/User"
                }
              }
            },
            "description" : "User created successfully"
          },
          "400" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/error"
                }
              }
            },
            "description" : "Bad Request"
          },
          "default" : {
            "description" : "Unexpected error"
          }
        },
        "summary" : "Create User",
        "tags" : [ "User" ]
      },
      "put" : {
        "operationId" : "updateUser",
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/User"
              }
            },
            "application/xml" : {
              "schema" : {
                "$ref" : "#/components/schemas/User"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/User"
                }
              }
            },
            "description" : "User updated successfully"
          },
          "400" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/error"
                }
              }
            },
            "description" : "Bad Request"
          },
          "404" : {
            "description" : "User not found"
          },
          "default" : {
            "description" : "Unexpected error"
          }
        },
        "summary" : "Update User",
        "tags" : [ "User" ]
      }
    },
    "/Users/{id}" : {
      "get" : {
        "operationId" : "findUserById",
        "parameters" : [ {
          "description" : "Status values that need to be considered for filter",
          "explode" : false,
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "format" : "int64",
            "type" : "integer"
          },
          "style" : "simple"
        } ],
        "responses" : {
          "200" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/User"
                }
              },
              "application/xml" : {
                "schema" : {
                  "items" : {
                    "$ref" : "#/components/schemas/User"
                  },
                  "type" : "array"
                }
              }
            },
            "description" : "successful operation"
          },
          "400" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/error"
                }
              }
            },
            "description" : "Invalid status value"
          },
          "404" : {
            "description" : "User not found"
          },
          "default" : {
            "description" : "Unexpected error"
          }
        },
        "summary" : "Finds User by Id..",
        "tags" : [ "User" ]
      }
    },
    "/Users/{id}/deleteUser" : {
      "put" : {
        "operationId" : "deleteUser",
        "parameters" : [ {
          "description" : "ID of user to delete",
          "explode" : false,
          "in" : "path",
          "name" : "id",
          "required" : true,
          "schema" : {
            "format" : "int64",
            "type" : "integer"
          },
          "style" : "simple"
        } ],
        "responses" : {
          "204" : {
            "description" : "User deleted successfully"
          },
          "400" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/error"
                }
              }
            },
            "description" : "Bad Request"
          },
          "404" : {
            "description" : "User not found"
          },
          "default" : {
            "description" : "Unexpected error"
          }
        },
        "summary" : "Delete User",
        "tags" : [ "User" ]
      }
    },
...
  },
  "components" : {
    "schemas" : {
      "paging" : {
        "example" : {
          "pageNumber" : 1,
          "pageIndex" : 1,
          "pageSize" : 10,
          "totalCount" : 100
        },
        "properties" : {
          "pageIndex" : {
            "example" : 1,
            "format" : "int64",
            "type" : "integer"
          },
          "pageSize" : {
            "example" : 10,
            "format" : "int64",
            "type" : "integer"
          },
          "totalCount" : {
            "example" : 100,
            "format" : "int64",
            "type" : "integer"
          },
          "pageNumber" : {
            "example" : 1,
            "format" : "int64",
            "type" : "integer"
          }
        },
        "type" : "object"
      },
      "User" : {
        "example" : {
          "firstName" : "John",
          "lastName" : "Doe",
          "role" : "TSE",
          "id" : 1,
          "userName" : "JohnDoe@123",
          "email" : "johnDoe@gmail.com"
        },
        "properties" : {
          "id" : {
            "example" : 1,
            "type" : "integer"
          },
          "firstName" : {
            "example" : "John",
            "type" : "string"
          },
          "lastName" : {
            "example" : "Doe",
            "type" : "string"
          },
          "userName" : {
            "example" : "JohnDoe@123",
            "type" : "string"
          },
          "email" : {
            "example" : "johnDoe@gmail.com",
            "format" : "email",
            "type" : "string"
          },
          "role" : {
            "example" : "TSE",
            "type" : "string"
          }
        },
        "required" : [ "email", "firstName", "lastName", "userName" ],
        "type" : "object"
      },
...
      "error" : {
        "example" : {
          "code" : 400,
          "message" : "Bad Request"
        },
        "properties" : {
          "code" : {
            "example" : 400,
            "type" : "integer"
          },
          "message" : {
            "example" : "Bad Request",
            "type" : "string"
          }
        },
        "required" : [ "code", "message" ],
        "type" : "object"
      },
      "getUsers_200_response" : {
        "example" : {
          "paging" : {
            "pageNumber" : 1,
            "pageIndex" : 1,
            "pageSize" : 10,
            "totalCount" : 100
          },
          "users" : [ {
            "firstName" : "John",
            "lastName" : "Doe",
            "role" : "TSE",
            "id" : 1,
            "userName" : "JohnDoe@123",
            "email" : "johnDoe@gmail.com"
          }, {
            "firstName" : "John",
            "lastName" : "Doe",
            "role" : "TSE",
            "id" : 1,
            "userName" : "JohnDoe@123",
            "email" : "johnDoe@gmail.com"
          } ]
        },
        "properties" : {
          "paging" : {
            "$ref" : "#/components/schemas/paging"
          },
          "users" : {
            "items" : {
              "$ref" : "#/components/schemas/User"
            },
            "type" : "array"
          }
        },
        "type" : "object"
      },
...
    }
  }
}